feat: further refactor of folder structure
