+++ src/components/Activity/components/index.tsx
@@ -1,0 +1,4 @@
+export { default as ChartAll } from "./ChartAll";
+export { default as ChartTop } from "./ChartTop";
+export { default as ChartYear } from "./ChartYear";
+export { ChartVizGenre, ChartVizRating, ChartVizReleased } from "./ChartViz";
