+++ src/components/Activity/fuego/fuegoChartActions.ts
@@ -1,6 +1,6 @@
-import type { MediaDiaryWithId, MediaType } from "../types/typesMedia";
-import type fuego from "./fuego";
-import { fuegoDb } from "./fuego";
+import type { MediaDiaryWithId, MediaType } from "../../../types/typesMedia";
+import type fuego from "../../../fuego/fuego";
+import { fuegoDb } from "../../../fuego/fuego";
 
 export async function fuegoChartYear(
   key: string,
