+++ src/components/Activity/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./Activity";
