+++ src/components/Settings/Settings.tsx
@@ -12,11 +12,11 @@ import {
 } from "@chakra-ui/react";
 import dayjs from "dayjs";
 import React, { useState } from "react";
-import { useMDState } from "../config/store";
-import useFuegoDelete from "../fuego/useFuegoDelete";
-import type { UserFuegoValidated } from "../types/typesUser";
-import LayoutModal from "./layouts/LayoutModal";
-import UserPreference from "./user/UserPreference";
+import { useMDState } from "../../config/store";
+import useFuegoDelete from "./fuego/useFuegoDelete";
+import type { UserFuegoValidated } from "../../types/typesUser";
+import LayoutModal from "../layouts/LayoutModal";
+import UserPreference from "../user/UserPreference";
 
 function Settings({ user }: { user: UserFuegoValidated }): JSX.Element {
   const { isOpen, onOpen, onClose } = useDisclosure();
