+++ src/components/icons/index.tsx
@@ -1,0 +1,14 @@
+export { default as ActivityIcon } from "./ActivityIcon";
+export { default as AlbumIcon } from "./AlbumIcon";
+export { default as BookmarkIcon } from "./BookmarkIcon";
+export { default as FilmIcon } from "./FilmIcon";
+export { default as FiltersIcon } from "./FiltersIcon";
+export { default as GoogleIcon } from "./GoogleIcon";
+export { default as HomeIcon } from "./HomeIcon";
+export { default as LayersIcon } from "./LayersIcon";
+export { default as LogoIcon } from "./LogoIcon";
+export { default as MemoryIcon } from "./MemoryIcon";
+export { default as PlusIcon } from "./PlusIcon";
+export { default as StartEmptyIcon } from "./StartEmptyIcon";
+export { default as TvIcon } from "./TvIcon";
+export { default as TwitterIcon } from "./TwitterIcon";
